backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
# LangChain Core and Integrations
langchain
langchain-google-genai 